+++ src/contents/ui/ClickableLabel.qml
@@ -1,14 +1,14 @@
 /**
- * Weekday Grid widget for KDE
+ * Weekday Grid widget for Plasma 6 / KDE
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2020-2023 Marcin Orlowski
+ * @copyright 2020-2026 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/weekday-plasmoid
  */
 
-import QtQuick 2.0
-import org.kde.plasma.components 3.0 as PlasmaComponents
+import QtQuick
+import org.kde.plasma.components as PlasmaComponents
 
 PlasmaComponents.Label {
 	// URL to open once the label is clicked.
@@ -29,4 +29,3 @@ PlasmaComponents.Label {
 		}
 	}
 }
-