+++ src/contents/ui/FakeCalendarModeWarning.qml
@@ -1,16 +1,17 @@
 /**
- * Weekday Grid widget for KDE
+ * Weekday Grid widget for Plasma 6 / KDE
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2020-2023 Marcin Orlowski
+ * @copyright 2020-2026 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/weekday-plasmoid
  */
 
-import QtQuick 2.0
-import QtQuick.Layouts 1.1
-import org.kde.kirigami 2.5 as Kirigami
-import org.kde.plasma.components 3.0 as PlasmaComponents
+import QtQuick
+import QtQuick.Layouts
+import org.kde.kirigami as Kirigami
+import org.kde.plasma.components as PlasmaComponents
+import org.kde.plasma.plasmoid
 
 Kirigami.InlineMessage {
 	Layout.fillWidth: true
@@ -18,6 +19,5 @@ Kirigami.InlineMessage {
 	type: Kirigami.MessageType.Warning
 	text: 'Widget "Fake parameters" mode is enabled in "Theme" editor.'
 	showCloseButton: true
-	visible: plasmoid.configuration.useFakeParameters
+	visible: Plasmoid.configuration.useFakeParameters
 }
-