+++ src/contents/ui/Theme.qml
@@ -1,13 +1,13 @@
 /**
- * Weekday Grid widget for KDE
+ * Weekday Grid widget for Plasma 6 / KDE
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2020-2023 Marcin Orlowski
+ * @copyright 2020-2026 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/weekday-plasmoid
  */
 
-import QtQuick 2.0
+import QtQuick
 
 Item {
 	readonly property string defaultBg: '#00000000'
@@ -140,7 +140,7 @@ Item {
 				'italic': pastSundayItalic,
 			}
 		})
-	
+
 		json['today'] = {
 			'fg': todayFg,
 			'bg': todayBg,